Refactor symmetry classes
