+++ jaxmat/tensors/mappings.py
@@ -201,60 +201,3 @@ def kelvin_rank4_map(d: int) -> Tuple[IndexMap4, jax.Array]:
     W = W2[a] * W2[b]
 
     return (I, J, K, L), W
-
-
-def isotropic_projectors(d=3):
-    """Isotropic projectors J and K in (6, 6) Kelvin-Mandel format."""
-    I = jnp.eye(d)
-    I4 = jnp.einsum("ij,kl->ijkl", I, I)
-    I4s = 0.5 * (jnp.einsum("ik,jl->ijkl", I, I) + jnp.einsum("il,jk->ijkl", I, I))
-
-    J4 = (1.0 / d) * I4
-    K4 = I4s - J4
-
-    (i, j, k, l), W = kelvin_rank4_map(d)
-
-    J = W * J4[i, j, k, l]
-    K = W * K4[i, j, k, l]
-
-    return J, K, J4, K4
-
-
-def cubic_projectors(d=3):
-    """
-    Cubic symmetry projectors in 6x6 Kelvin-Mandel format.
-
-    Returns
-    -------
-    J, Ka, Kb : jax.Array
-        6x6 Kelvin-Mandel projectors
-    J4, Ka4, Kb4 : jax.Array
-        4th-rank tensor projectors (d,d,d,d)
-    """
-    I = jnp.eye(d)
-    I4 = jnp.einsum("ij,kl->ijkl", I, I)
-    I4s = 0.5 * (jnp.einsum("ik,jl->ijkl", I, I) + jnp.einsum("il,jk->ijkl", I, I))
-
-    # 1. Spherical projector
-    J4 = (1.0 / d) * I4
-
-    # 2. Cubic invariant tensor
-    Lambda = jnp.zeros((d, d, d, d))
-    for i in range(d):
-        Lambda = Lambda.at[i, i, i, i].set(1.0)
-
-    # 3. Cubic decomposition of deviatoric part
-    # Ka = Lambda - J = deviatoric projector of diagonal part
-    Ka4 = Lambda - J4
-
-    # Remaining part is off-diagonal shear projector
-    Kb4 = I4s - Lambda
-
-    # Kelvin mapping
-    (i, j, k, l), W = kelvin_rank4_map(d)
-
-    J = W * J4[i, j, k, l]
-    Ka = W * Ka4[i, j, k, l]
-    Kb = W * Kb4[i, j, k, l]
-
-    return J, Ka, Kb, J4, Ka4, Kb4
